Split up into separate files Add arrivalTime and arrivalTimeRelative as result to .monitor(), see GH issue #3
