+++ index.js
@@ -1,3 +1,9 @@
 'use strict';
 
-module.exports = require('./lib/dvbjs');
+var dvb = {
+    find: require('./lib/find'),
+    route: require('./lib/route'),
+    monitor: require('./lib/monitor')
+};
+
+module.exports = dvb;
